+++ desktop/src/main.rs
@@ -2,7 +2,11 @@
 
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // disables console spawning for release build
 
+use anyhow::bail;
+use fd_lock::RwLock;
 use reqwest::Client;
+use std::fs::{File, OpenOptions};
+use std::io::prelude::*;
 
 #[macro_use]
 extern crate rust_i18n;
@@ -32,7 +36,28 @@ pub fn get_reqwest_client() -> &'static reqwest::Client {
 }
 
 fn main() -> anyhow::Result<()> {
-    env_logger::init();
+    let mut p = dirs::config_dir().expect("failed to find config directory");
+    p.push("JukeBoxDesktop");
+    p.push("app.lock");
+
+    let mut f = RwLock::new(
+        OpenOptions::new()
+            .read(true)
+            .write(true)
+            .create(true)
+            .open(p)
+            .unwrap(),
+    );
+    let f = f.try_write();
+    if let Err(_) = f {
+        bail!("failed to acquire exclusive lock for application. aborting.")
+    }
+
+    // env_logger::init();
+    let mut p = dirs::config_dir().expect("failed to find config directory");
+    p.push("JukeBoxDesktop");
+    p.push("jukebox_desktop_debug.log");
+    simple_logging::log_to_file(p, log::LevelFilter::Debug).unwrap();
 
     // For OBS websocket TLS support, currently unused.
     // rustls::crypto::aws_lc_rs::default_provider()
@@ -41,6 +66,8 @@ fn main() -> anyhow::Result<()> {
 
     // GUI launches all the necessary threads when started
     gui::gui::basic_gui();
+
+    drop(f);
 
     Ok(())
 }